# PERF: improve MultiIndex get_loc performance