--- a/pandas/tests/util/test_hashing.py
+++ b/pandas/tests/util/test_hashing.py
@@ -6,7 +6,7 @@ import pandas as pd
 
 from pandas import DataFrame, Series, Index, MultiIndex
 from pandas.util import hash_array, hash_pandas_object
-from pandas.core.util.hashing import hash_tuples, hash_tuple
+from pandas.core.util.hashing import hash_tuples, hash_tuple, _hash_scalar
 import pandas.util.testing as tm
 
 
@@ -81,10 +81,20 @@ class TestHashing(object):
 
     def test_hash_tuple(self):
         # test equivalence between hash_tuples and hash_tuple
-        tup = (1, 'one')
-        result = hash_tuple(tup)
-        expected = hash_tuples([tup])[0]
-        assert result == expected
+        for tup in [(1, 'one'), (1, np.nan)]:
+            result = hash_tuple(tup)
+            expected = hash_tuples([tup])[0]
+            assert result == expected
+
+    def test_hash_scalar(self):
+        for val in [1, 1.4, 'A', b'A', u'A',  pd.Timestamp("2012-01-01"),
+                    pd.Timestamp("2012-01-01", tz='Europe/Brussels'),
+                    pd.Period('2012-01-01', freq='D'), pd.Timedelta('1 days'),
+                    pd.Interval(0, 1), np.nan, pd.NaT, None]:
+            result = _hash_scalar(val)
+            expected = hash_array(np.array([val], dtype=object),
+                                  categorize=True)
+            assert result[0] == expected[0]
 
     def test_hash_tuples_err(self):