@jreback 2017-05-16
@jorisvandenbossche lgtm. merge when ready (for 0.20.2) is good.