@jreback 2017-05-17
nice fix @jorisvandenbossche !